Refatorar FilialController

+++ Service/Filial/FilialService.cs
@@ -1,6 +1,8 @@
+using System.Threading.Tasks;
 using CollectApp.Data;
 using CollectApp.Models;
 using CollectApp.ViewModels;
+using Microsoft.CodeAnalysis.Differencing;
 using Microsoft.EntityFrameworkCore;
 
 namespace CollectApp.Services
@@ -14,8 +16,13 @@ namespace CollectApp.Services
             _context = context;
         }
 
-        public async Task<OperationResult> AddFilial(Filial filial)
+        public async Task<OperationResult> CreateFilial(CreateFilialViewModel filialCreate)
         {
+            Filial filial = new Filial
+            {
+                Name = filialCreate.Name,
+            };
+
             bool productExist = await _context.Filials.AnyAsync(f => f.Name == filial.Name && f.Id != filial.Id);
 
             if (productExist)
@@ -24,18 +31,49 @@ namespace CollectApp.Services
             }
 
             _context.Filials.Add(filial);
+            await SaveChangesFilialsAsync();
 
             return OperationResult.Ok();
         }
 
+        public async Task<EditFilialViewModel> SetEditFilialViewModel(int? id)
+        {
+            Filial? filial = await FindFilialAsync(id);
+
+            if (filial == null)
+            {
+                EditFilialViewModel NotFound = new EditFilialViewModel();
+                return NotFound;
+            }
+
+            EditFilialViewModel epvm = new EditFilialViewModel
+            {
+                Id = filial.Id,
+                Name = filial.Name,
+            };
+
+            return epvm;
+        }
+
         public async Task<OperationResult> EditFilial(EditFilialViewModel filialEdit)
         {
-            bool productExist = await _context.Filials.AnyAsync(f => f.Name == filialEdit.Name && f.Id != filialEdit.Id);
+            Filial? filial = await FindFilialAsync(filialEdit.Id);
 
-            if (productExist)   
+            if (filial == null)
+            {
+                OperationResult NotFound = new OperationResult();
+                return NotFound;
+            }
+
+            bool filialExist = await _context.Filials.AnyAsync(f => f.Name == filialEdit.Name && f.Id != filialEdit.Id);
+
+            if (filialExist)
             {
                 return OperationResult.Fail($"Já existe uma filial cadastrada com o nome fornecido");
             }
+
+            filial.Name = filialEdit.Name;
+            await SaveChangesFilialsAsync();
 
             return OperationResult.Ok();
         }
@@ -55,14 +93,40 @@ namespace CollectApp.Services
             return await _context.Filials.Where(f => f.Name.Contains(input)).ToListAsync();
         }
 
+        public async Task<List<FilialListViewModel>> SetFilialListViewModel()
+        {
+            List<Filial> filials = await GetAllFilialsListAsycn();
+
+            List<FilialListViewModel> flvm = filials.Select(f => new FilialListViewModel
+            {
+                Id = f.Id,
+                Name = f.Name
+            }).ToList();
+
+            return flvm;
+        }
+
         public async Task<int> SaveChangesFilialsAsync()
         {
             return await _context.SaveChangesAsync();
         }
 
-        public void DeleteFilial(Filial filial)
+        public async Task DeleteFilial(int? id)
         {
+            if (id == null)
+            {
+                return;
+            }
+
+            Filial? filial = await FindFilialAsync(id);
+
+            if (filial == null)
+            {
+                return;
+            }
+
             _context.Filials.Remove(filial);
+            await SaveChangesFilialsAsync();
         }
     }
 }
